+++ src/ui/map.test.ts
@@ -1818,6 +1818,31 @@ describe('Map', () => {
             });
         });
 
+        test('#3373 paint property should be synchronized with an update', async () => {
+            const colors = ['red', 'blue'];
+            const map = createMap({
+                style: {
+                    'version': 8,
+                    'sources': {},
+                    'layers': [{
+                        'id': 'background',
+                        'type': 'background',
+                        'paint': {
+                            'background-color': colors[0]
+                        }
+                    }]
+                }
+            });
+
+            await map.once('style.load');
+            expect(map.getPaintProperty('background', 'background-color')).toBe(colors[0]);
+            expect(map.getStyle().layers.filter(l => l.id === 'background')[0].paint['background-color']).toBe(colors[0]);
+            // update property
+            map.setPaintProperty('background', 'background-color', colors[1]);
+            expect(map.getPaintProperty('background', 'background-color')).toBe(colors[1]);
+            expect(map.getStyle().layers.filter(l => l.id === 'background')[0].paint['background-color']).toBe(colors[1]);
+        });
+
         test('throw before loaded', () => {
             const map = createMap({
                 style: {